requirements.txt
# Agent Swarm Framework Requirements
openai>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
tenacity>=8.0.0
typing-extensions>=4.0.0
colorama>=0.4.6
pyyaml>=6.0.0
jsonschema>=4.0.0
flask>=3.0.0
flask-cors>=4.0.0